requirements.txt
Flask
gunicorn
openai
PyYAML
requests
